+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [1.2.1] - 2026-05-18
+## [1.3.0] - 2026-06-01
 
-### Fix
-- Fixed an issue with chunk size for server summarization as it was pushing past the token limit for provider. Also made changes to server summary prompt to be more concise and utilize less tokens on the prompt to get a more user friendly summarization and more readable output that is more accurate to the summarization.
+### Sticky Messages
+Added the ability for admins to use /sticky commands to pin messages to the bottom of channels as well as remove them at will and see all their sticky messages using /sticky view.